alterando tipo para deploy vercel

+++ cp6/src/app/api/base-challenge/[id]/route.ts
@@ -1,19 +1,20 @@
 import { TipoChallenge } from "@/types";
-import { NextResponse } from "next/server";
+import { NextResponse, NextRequest } from "next/server";
 import { client } from "@/lib/appwrite_client";
 import { Databases } from "appwrite";
 
 const database = new Databases(client);
 
-export async function GET(
-  request: Request,
-  { params }: { params: { id: string } }
-) {
+export async function GET(request: NextRequest) {
+  const id = request.nextUrl.searchParams.get('id');
+  if (!id) {
+    return NextResponse.json({ message: "ID não fornecido" }, { status: 400 });
+  }
   try {
     const response = await database.getDocument(
       process.env.NEXT_PUBLIC_APPWRITE_DATABASE_ID as string,
       process.env.NEXT_PUBLIC_APPWRITE_COLLECTION_CHALLENGE_ID as string,
-      params.id
+      id
     );
     return NextResponse.json(response);
   } catch (error) {
@@ -25,13 +26,13 @@ export async function GET(
   }
 }
 
-export async function PUT(
-  request: Request,
-  { params }: { params: { id: string } }
-) {
+export async function PUT(request: NextRequest) {
+  const id = request.nextUrl.searchParams.get('id');
+  if (!id) {
+    return NextResponse.json({ message: "ID não fornecido" }, { status: 400 });
+  }
   try {
-    const { nome, descricao, nota } =
-      await request.json();
+    const { nome, descricao, nota } = await request.json();
     const challenge = {
       nome,
       descricao,
@@ -41,7 +42,7 @@ export async function PUT(
     await database.updateDocument(
       process.env.NEXT_PUBLIC_APPWRITE_DATABASE_ID as string,
       process.env.NEXT_PUBLIC_APPWRITE_COLLECTION_CHALLENGE_ID as string,
-      params.id,
+      id,
       challenge
     );
 
@@ -55,15 +56,16 @@ export async function PUT(
   }
 }
 
-export async function DELETE(
-  request: Request,
-  { params }: { params: { id: string } }
-) {
+export async function DELETE(request: NextRequest) {
+  const id = request.nextUrl.searchParams.get('id');
+  if (!id) {
+    return NextResponse.json({ message: "ID não fornecido" }, { status: 400 });
+  }
   try {
     await database.deleteDocument(
       process.env.NEXT_PUBLIC_APPWRITE_DATABASE_ID as string,
       process.env.NEXT_PUBLIC_APPWRITE_COLLECTION_CHALLENGE_ID as string,
-      params.id
+      id
     );
 
     return NextResponse.json({ status: 204 });
